requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.1
Unidecode==0.04.17
blinker==1.3
bottlenose==0.6.2
docutils==0.12
feedgenerator==1.7
lxml==3.4.1
pelican==3.5.0
python-amazon-simple-product-api==2.0.0
python-dateutil==2.4.0
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
